Changes float for double

+++ ComPay/src/main/java/com/tsw/ComPay/Dto/GroupDto.java
@@ -10,6 +10,6 @@ public class GroupDto {
     private Long id;
     private String groupName;
     private String currency;
-    private float amount;
+    private double amount;
     private String imgURL;
 }
